REF: Rfc fragments to Compose screens, navigation and DI
Rfc fragments and activity to compose only.
Rfc navigation graph to compose, create Screen class.

Remove unused and add new compose dependencies
