+++ app/src/main/java/cz/quanti/razym/rocketapp/system/RocketLaunchFragment.kt
@@ -1,9 +1,7 @@
+@file:OptIn(ExperimentalMaterial3Api::class)
+
 package cz.quanti.razym.rocketapp.system
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -18,81 +16,61 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
 import cz.quanti.razym.rocketapp.R
-import cz.quanti.razym.rocketapp.ui.theme.RocketappTheme
 
-@OptIn(ExperimentalMaterial3Api::class)
-class RocketLaunchFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        return ComposeView(requireContext()).apply {
-            setViewCompositionStrategy(
-                ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed
+@Preview
+@Composable
+fun RocketLaunchScreen(
+    onBackClick: () -> Unit = {},
+) {
+    Scaffold(
+        topBar = {
+            RocketLaunchFragmentTopBar(
+                title = stringResource(R.string.launch),
+                onBackClick = onBackClick,
             )
-            setContent {
-                RocketappTheme {
-                    RocketLaunchFragmentContent()
-                }
-            }
+        }
+    ) { innerPadding ->
+        Box(modifier = Modifier
+            .fillMaxSize()
+            .padding(innerPadding)
+        ) {
+            Text(
+                modifier = Modifier
+                    .align(Alignment.Center),
+                text = "Rocket Launch",
+                style = MaterialTheme.typography.titleMedium,
+            )
         }
     }
+}
 
-    @Preview
-    @Composable
-    private fun RocketLaunchFragmentContent() {
-        Scaffold(
-            topBar = {
-                RocketLaunchFragmentTopBar(stringResource(R.string.launch))
-            }
-        ) { innerPadding ->
-            Box(modifier = Modifier
-                .fillMaxSize()
-                .padding(innerPadding)
-            ) {
-                Text(
-                    modifier = Modifier
-                        .align(Alignment.Center),
-                    text = "Rocket Launch",
-                    style = MaterialTheme.typography.titleMedium,
-                )
-            }
-        }
-    }
-
-    @Composable
-    private fun RocketLaunchFragmentTopBar(
-        title: String,
-    ) {
-        CenterAlignedTopAppBar(
-            title = {
-                Text(
-                    text = title,
-                    style = MaterialTheme.typography.titleLarge,
-                )
-            },
-            navigationIcon = {
-                Icon(
-                    modifier = Modifier
-                        .size(36.dp)
-                        .padding(8.dp)
-                        .clickable {
-                            findNavController().navigateUp()
-                        },
-                    painter = painterResource(id = R.drawable.ic_arrow_back),
-                    // TODO
-                    contentDescription = "",
-                )
-            },
-        )
-    }
+@Composable
+private fun RocketLaunchFragmentTopBar(
+    title: String,
+    onBackClick: () -> Unit = {},
+) {
+    CenterAlignedTopAppBar(
+        title = {
+            Text(
+                text = title,
+                style = MaterialTheme.typography.titleLarge,
+            )
+        },
+        navigationIcon = {
+            Icon(
+                modifier = Modifier
+                    .size(36.dp)
+                    .padding(8.dp)
+                    .clickable { onBackClick() },
+                painter = painterResource(id = R.drawable.ic_arrow_back),
+                // TODO
+                contentDescription = "",
+            )
+        },
+    )
 }
