--- app/src/main/res/navigation/nav_graph.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nav_graph"
-    app:startDestination="@id/rocket_list_dest">
-
-    <fragment
-        android:id="@+id/rocket_list_dest"
-        android:name="cz.quanti.razym.rocketapp.system.RocketListFragment"
-        android:label="fragment_rocket_list"
-        >
-        <action
-            android:id="@+id/action_rocketListFragment_to_rocketDetailFragment"
-            app:destination="@id/rocket_detail_dest" />
-    </fragment>
-    <fragment
-        android:id="@+id/rocket_detail_dest"
-        android:name="cz.quanti.razym.rocketapp.system.RocketDetailFragment"
-        android:label="fragment_rocket_detail"
-        >
-        <action
-            android:id="@+id/action_rocketDetailFragment_to_rocketLaunchFragment"
-            app:destination="@id/rocket_launch_dest" />
-        <!--suppress NavigationFile (rocketId not a valid destination) -->
-        <argument
-            android:name="rocketId"
-            app:argType="string" />
-        <!--suppress NavigationFile -->
-        <argument
-            android:name="rocketName"
-            app:argType="string" />
-    </fragment>
-    <fragment
-        android:id="@+id/rocket_launch_dest"
-        android:name="cz.quanti.razym.rocketapp.system.RocketLaunchFragment"
-        android:label="fragment_rocket_launch"
-        />
-
-</navigation>
