+++ build.gradle.kts
@@ -3,14 +3,3 @@ plugins {
     id("com.android.application") version "8.1.3" apply false
     id("org.jetbrains.kotlin.android") version "1.9.20" apply false
 }
-
-buildscript {
-    val navVersion by rootProject.extra { "2.7.5" }
-
-    repositories {
-        google()
-    }
-    dependencies {
-        classpath("androidx.navigation:navigation-safe-args-gradle-plugin:$navVersion")
-    }
-}
